backend/requirements.txt
Flask==3.0.3
flask_cors
openai
langchain
langchain-openai
langchain-community
langchain-core
langchainhub
chromadb
python-dotenv
beautifulsoup4
selenium
webdriver-manager
